refactor: 공연 domian/entity  common 패키지로 분리
