+++ src/main/java/com/prography/lighton/common/exception/GlobalExceptionHandler.java
@@ -8,9 +8,9 @@ import com.prography.lighton.common.exception.base.NotFoundException;
 import com.prography.lighton.common.exception.base.UnsupportedTypeException;
 import com.prography.lighton.common.utils.ApiUtils;
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
-import com.prography.lighton.performance.users.domain.exception.MasterArtistCannotBeRemovedException;
-import com.prography.lighton.performance.users.domain.exception.PerformanceNotApprovedException;
-import com.prography.lighton.performance.users.domain.exception.PerformanceUpdateNotAllowedException;
+import com.prography.lighton.performance.common.domain.exception.MasterArtistCannotBeRemovedException;
+import com.prography.lighton.performance.common.domain.exception.PerformanceNotApprovedException;
+import com.prography.lighton.performance.common.domain.exception.PerformanceUpdateNotAllowedException;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
