+++ src/main/java/com/prography/lighton/performance/admin/application/PendingPerformanceQueryUseCase.java
@@ -2,7 +2,7 @@ package com.prography.lighton.performance.admin.application;
 
 import com.prography.lighton.performance.admin.presentation.dto.response.GetPerformanceApplicationDetailResponseDTO;
 import com.prography.lighton.performance.admin.presentation.dto.response.GetPerformanceApplicationListResponseDTO;
-import com.prography.lighton.performance.users.domain.entity.enums.ApproveStatus;
+import com.prography.lighton.performance.common.domain.entity.enums.ApproveStatus;
 
 public interface PendingPerformanceQueryUseCase {
