+++ src/main/java/com/prography/lighton/performance/admin/infrastructure/repository/AdminPerformanceRepository.java
@@ -1,7 +1,7 @@
 package com.prography.lighton.performance.admin.infrastructure.repository;
 
-import com.prography.lighton.performance.users.domain.entity.Performance;
-import com.prography.lighton.performance.users.domain.entity.enums.ApproveStatus;
+import com.prography.lighton.performance.common.domain.entity.Performance;
+import com.prography.lighton.performance.common.domain.entity.enums.ApproveStatus;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
