+++ src/main/java/com/prography/lighton/performance/common/domain/entity/Performance.java
@@ -1,20 +1,20 @@
-package com.prography.lighton.performance.users.domain.entity;
+package com.prography.lighton.performance.common.domain.entity;
 
 import com.prography.lighton.artist.common.domain.entity.Artist;
 import com.prography.lighton.artist.common.domain.entity.exception.NotAMasterArtistException;
 import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.genre.domain.entity.Genre;
-import com.prography.lighton.performance.users.domain.entity.association.PerformanceArtist;
-import com.prography.lighton.performance.users.domain.entity.enums.ApproveStatus;
-import com.prography.lighton.performance.users.domain.entity.enums.Seat;
-import com.prography.lighton.performance.users.domain.entity.enums.Type;
-import com.prography.lighton.performance.users.domain.entity.vo.Info;
-import com.prography.lighton.performance.users.domain.entity.vo.Location;
-import com.prography.lighton.performance.users.domain.entity.vo.Payment;
-import com.prography.lighton.performance.users.domain.entity.vo.Schedule;
-import com.prography.lighton.performance.users.domain.exception.MasterArtistCannotBeRemovedException;
-import com.prography.lighton.performance.users.domain.exception.PerformanceNotApprovedException;
-import com.prography.lighton.performance.users.domain.exception.PerformanceUpdateNotAllowedException;
+import com.prography.lighton.performance.common.domain.entity.association.PerformanceArtist;
+import com.prography.lighton.performance.common.domain.entity.enums.ApproveStatus;
+import com.prography.lighton.performance.common.domain.entity.enums.Seat;
+import com.prography.lighton.performance.common.domain.entity.enums.Type;
+import com.prography.lighton.performance.common.domain.entity.vo.Info;
+import com.prography.lighton.performance.common.domain.entity.vo.Location;
+import com.prography.lighton.performance.common.domain.entity.vo.Payment;
+import com.prography.lighton.performance.common.domain.entity.vo.Schedule;
+import com.prography.lighton.performance.common.domain.exception.MasterArtistCannotBeRemovedException;
+import com.prography.lighton.performance.common.domain.exception.PerformanceNotApprovedException;
+import com.prography.lighton.performance.common.domain.exception.PerformanceUpdateNotAllowedException;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.CollectionTable;
 import jakarta.persistence.Column;
