+++ src/main/java/com/prography/lighton/performance/common/domain/entity/PerformanceGenre.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.performance.users.domain.entity;
+package com.prography.lighton.performance.common.domain.entity;
 
 import com.prography.lighton.common.domain.BaseEntity;
 import com.prography.lighton.genre.domain.entity.Genre;
