+++ src/main/java/com/prography/lighton/performance/common/domain/entity/enums/Seat.java
@@ -1,0 +1,5 @@
+package com.prography.lighton.performance.common.domain.entity.enums;
+
+public enum Seat {
+    STANDING, FREESTYLE, ASSIGNED
+}
