+++ src/main/java/com/prography/lighton/performance/common/domain/entity/enums/Type.java
@@ -1,0 +1,5 @@
+package com.prography.lighton.performance.common.domain.entity.enums;
+
+public enum Type {
+    NORMAL, BUSKING
+}
