+++ src/main/java/com/prography/lighton/performance/common/domain/entity/vo/Info.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.performance.users.domain.entity.vo;
+package com.prography.lighton.performance.common.domain.entity.vo;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
