+++ src/main/java/com/prography/lighton/performance/common/domain/exception/MasterArtistCannotBeRemovedException.java
@@ -1,10 +1,10 @@
-package com.prography.lighton.performance.users.domain.exception;
+package com.prography.lighton.performance.common.domain.exception;
 
 import com.prography.lighton.common.utils.ApiUtils;
 import org.springframework.http.HttpStatus;
 
-public class PerformanceUpdateNotAllowedException extends RuntimeException {
-    private static final String message = "공연 시작일 3일 전까지만 수정이 가능합니다.";
+public class MasterArtistCannotBeRemovedException extends RuntimeException {
+    private static final String message = "공연 등록 아티스트는 삭제할 수 없습니다.";
 
     public ApiUtils.ApiResult<?> body() {
         return ApiUtils.error(HttpStatus.CONFLICT, message);
