+++ src/main/java/com/prography/lighton/performance/users/application/resolver/PerformanceResolver.java
@@ -5,11 +5,11 @@ import com.prography.lighton.artist.users.application.service.ArtistService;
 import com.prography.lighton.genre.application.service.GenreService;
 import com.prography.lighton.genre.domain.entity.Genre;
 import com.prography.lighton.member.domain.entity.Member;
-import com.prography.lighton.performance.users.domain.entity.enums.Seat;
-import com.prography.lighton.performance.users.domain.entity.vo.Info;
-import com.prography.lighton.performance.users.domain.entity.vo.Location;
-import com.prography.lighton.performance.users.domain.entity.vo.Payment;
-import com.prography.lighton.performance.users.domain.entity.vo.Schedule;
+import com.prography.lighton.performance.common.domain.entity.enums.Seat;
+import com.prography.lighton.performance.common.domain.entity.vo.Info;
+import com.prography.lighton.performance.common.domain.entity.vo.Location;
+import com.prography.lighton.performance.common.domain.entity.vo.Payment;
+import com.prography.lighton.performance.common.domain.entity.vo.Schedule;
 import com.prography.lighton.performance.users.presentation.dto.InfoDTO;
 import com.prography.lighton.performance.users.presentation.dto.PaymentDTO;
 import com.prography.lighton.performance.users.presentation.dto.ScheduleDTO;
