+++ src/main/java/com/prography/lighton/performance/users/application/service/PerformanceService.java
@@ -3,9 +3,9 @@ package com.prography.lighton.performance.users.application.service;
 import com.prography.lighton.artist.common.domain.entity.Artist;
 import com.prography.lighton.artist.users.application.service.ArtistService;
 import com.prography.lighton.member.domain.entity.Member;
+import com.prography.lighton.performance.common.domain.entity.Performance;
+import com.prography.lighton.performance.common.domain.entity.enums.Type;
 import com.prography.lighton.performance.users.application.resolver.PerformanceResolver;
-import com.prography.lighton.performance.users.domain.entity.Performance;
-import com.prography.lighton.performance.users.domain.entity.enums.Type;
 import com.prography.lighton.performance.users.infrastructure.repository.PerformanceRepository;
 import com.prography.lighton.performance.users.presentation.dto.PerformanceRegisterRequest;
 import com.prography.lighton.performance.users.presentation.dto.PerformanceUpdateRequest;
