--- src/main/java/com/prography/lighton/performance/users/domain/entity/enums/Seat.java
@@ -1,5 +1,0 @@
-package com.prography.lighton.performance.users.domain.entity.enums;
-
-public enum Seat {
-    STANDING, FREESTYLE, ASSIGNED
-}
