--- src/main/java/com/prography/lighton/performance/users/domain/entity/enums/Type.java
@@ -1,5 +1,0 @@
-package com.prography.lighton.performance.users.domain.entity.enums;
-
-public enum Type {
-    NORMAL, BUSKING
-}
