+++ src/main/java/com/prography/lighton/performance/users/presentation/exception/PerformanceExceptionHandler.java
@@ -1,8 +1,8 @@
 package com.prography.lighton.performance.users.presentation.exception;
 
 import com.prography.lighton.common.utils.ApiUtils.ApiResult;
-import com.prography.lighton.performance.users.domain.exception.MasterArtistCannotBeRemovedException;
-import com.prography.lighton.performance.users.domain.exception.PerformanceNotApprovedException;
+import com.prography.lighton.performance.common.domain.exception.MasterArtistCannotBeRemovedException;
+import com.prography.lighton.performance.common.domain.exception.PerformanceNotApprovedException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
